% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_xml_to_table.R
\name{clinical_xml_to_table}
\alias{clinical_xml_to_table}
\title{clinical_xml_to_table}
\usage{
clinical_xml_to_table(
  xml_data,
  data_items = c("Time/Day", "Time/Hour", "Time/Minute", "Time/Second", "ReadingStatus",
    "ErrorStatus", "Systolic", "Diastolic", "MeanArterialPressure", "HeartRate",
    "PulsePressure", "CoefficientOfVariability", "DiaryEventAnnotation", "Excluded"),
  data_types = c("i", "i", "i", "i", "c", "c", "i", "i", "i", "i", "i", "i", "c", "c"),
  parent_level = "//Measurements/Measurement/"
)
}
\arguments{
\item{xml_data}{this is the clinical_data list element produced from the read_rsntl() function}

\item{data_items}{this is a vector of items that are wanted the default}

\item{data_types}{this is a vector of data types for items that are wanted}

\item{parent_level}{this is the common parent level in the xml below which the dataitems are located}
}
\value{
A tibble with the columns representing the data items
}
\description{
convert xml to a table
}
\details{
This is probably heavily dependent on the version of the data file format and could need regular
ammending it. The dataitems and parentlevel parameters are an attempt to make this flexible but
possibly will not be sufficient.
}
\examples{
xml_data <- read_rsntl(system.file("extdata","2013_B_07-01-2021_13-49-43.RSNTL",package="sentinelreadr"))
tables <- lapply(xml_data$clinical_data,clinical_xml_to_table)

}
