% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_model_version.R
\name{get_data_model_version}
\alias{get_data_model_version}
\title{get_data_model_version}
\usage{
get_data_model_version(xmlData)
}
\arguments{
\item{xmlData}{}
}
\description{
get the data model version from the metadata secton
}
