% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.R
\name{process_file_clinical_details}
\alias{process_file_clinical_details}
\title{process the clinical details}
\usage{
process_file_clinical_details(fn, db = F)
}
\arguments{
\item{fn}{the filename}
}
\description{
process the clinical details from an RSNTL file and return a record set 1 record per reading
}
