% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.R
\name{process_file_patient_details}
\alias{process_file_patient_details}
\title{process the patient details}
\usage{
process_file_patient_details(fn, db = F)
}
\arguments{
\item{fn}{the filename}
}
\description{
process the patient details from an RSNTL file and return a single record
}
