% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsntl.R
\name{read_rsntl}
\alias{read_rsntl}
\title{read_rsntl}
\usage{
read_rsntl(filename)
}
\arguments{
\item{filename}{}
}
\value{
a list of xml elements for each xml sections of the file

 metadata = PackageMetadata

 testdetails = PackageTestDetails

 clincial_data_files = ClinicalDataItems

 clinical_data = ABP

 report = Report

where there are multiple xml elements in a section this will be a list of of list
}
\description{
read a sentinel file extract the xml sections from the binary file size data
}
\examples{
xml_data <- read_rsntl(system.file("extdata","2013_B_07-01-2021_13-49-43.RSNTL",package="sentinelreadr"))

}
