% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdetails_xml_to_table.R
\name{testdetails_xml_to_table}
\alias{testdetails_xml_to_table}
\title{patient test details to table}
\usage{
testdetails_xml_to_table(xmlData)
}
\arguments{
\item{xmlData}{the xml data to parse}
}
\description{
convert
}
