% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_table.R
\name{xml_to_table}
\alias{xml_to_table}
\title{xml_to_table}
\usage{
xml_to_table(xmlData, dataItems, dataTypes, parentLevel)
}
\arguments{
\item{xmlData}{this is the clinical_data list element produced from the read_rsntl() function}

\item{dataitems}{this is a vector of items that are wanted the default}

\item{parentlevel}{this is the common parent level in the xml below which the dataitems are located}
}
\value{
A data.table with the columns representing the data items
}
\description{
convert xml to a table
}
\details{
This is probably heavily dependent on the version of the data file format and could need regular
ammending it. The dataitems and parentlevel parameters are an attempt to make this flexible but
possibly will not be sufficient.
}
\examples{
xml_data <- read_rsntl(system.file("extdata","2013_B_07-01-2021_13-49-43.RSNTL",package="sentinelreadr"))
tables <- lapply(xml_data$clinical_data,clinical_xml_to_table)

}
