% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.R
\name{process_files}
\alias{process_files}
\title{Process a directory of files}
\usage{
process_files(
  directory,
  extention = "RSNTL",
  pattablefile = NULL,
  restablefile = NULL,
  silent = F,
  db = F
)
}
\description{
Produce two files one with patient detail one with all test results
}
